requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.8
django-allauth==0.58.2
django-bootstrap5==23.3
django-phonenumber-field==7.2.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.47.0
freetype-py==2.3.0
frozenlist==1.4.1
gunicorn==21.2.0
html5lib==1.1
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.4
mlxtend==0.23.0
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.24.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.2
pandas==2.0.3
phonenumbers==8.13.26
Pillow==10.1.0
psycopg2-binary==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf==3.17.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.8
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.3.0
scikit-learn==1.1.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
threadpoolctl==3.2.0
tinycss2==1.2.1
twilio==8.11.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.13
yarl==1.9.4
zipp==3.17.0
